Pencil configurations

A pencil configuration refers to a given distribution of multidimensional data among MPI processes. This information is encoded in the Pencil type.

A pencil configuration includes:

MPI topology information,

global and local dimensions of the numerical grid,

subset of decomposed dimensions,

definition of optional permutation of dimensions.

Construction

The creation of a new Pencil requires a MPITopology and the global data dimensions. One may also specify the list of decomposed dimensions, as well as an optional permutation of dimensions.

For instance, to decompose along 2 dimensions of a 3D dataset,

topology = MPITopology(comm, (8, 4))  # assuming 8×4 = 32 processes
dims_global = (16, 32, 64)
pencil = Pencil(topology, dims_global)

By default, the decomposed dimensions are the rightmost ones (in this case, dimensions 2 and 3). A different set of dimensions may be selected via the optional positional argument. For instance, to decompose along dimensions 1 and 3 instead,

decomp_dims = (1, 3)
pencil = Pencil(topology, dims_global, decomp_dims)

One may also want to work with multiple pencil configurations that differ, for instance, on the selection of decomposed dimensions. For this case, a second constructor is available that takes an already existing Pencil instance. Calling this constructor should be preferred when possible since it allows sharing memory buffers (used for instance for global transpositions) and thus reducing memory usage. The following creates a Pencil equivalent to the one above, but with different decomposed dimensions:

pencil_x = Pencil(pencil; decomp_dims=(1, 2))

See the Pencil documentation for more details.

Dimension permutations

As mentioned above, a Pencil may optionally be given information on dimension permutations. In this case, the layout of the data arrays in memory is different from the logical order of dimensions. For performance reasons, permutations are compile-time objects defined in the StaticPermutations package.

To make permutations clearer, consider the example above where the global data dimensions are $N_x × N_y × N_z = 16 × 32 × 64$. In this case, the logical order is $(x, y, z)$. Now let's say that we want the memory order of the data to be $(y, z, x)$,[1] which corresponds to the permutation (2, 3, 1).

Permutations are passed to the Pencil constructor via the permute keyword argument. Dimension permutations should be specified using a Permutation object. For instance,

permutation = Permutation(2, 3, 1)
pencil = Pencil(#= ... =#, permute=permutation)

One can also pass a NoPermutation to disable permutations (this is the default).

Types

PencilArrays.Pencils.Pencil — Type

Pencil{N,M}

Describes the decomposition of an N-dimensional array among MPI processes along M directions (with M < N).

Pencil(
    topology::MPITopology{M}, size_global::Dims{N},
    decomp_dims::Dims{M} = default_decomposition(N, Val(M));
    permute::AbstractPermutation = NoPermutation(),
    timer = TimerOutput(),
)

Define the decomposition of an N-dimensional geometry along M dimensions.

The dimensions of the geometry are given by size_global = (N1, N2, ...). The Pencil describes the decomposition of an array of dimensions size_global across a group of MPI processes.

Data is distributed over the given M-dimensional MPI topology (with M < N).

The decomposed dimensions may optionally be provided via the decomp_dims argument. By default, the M rightmost dimensions are decomposed. For instance, for a 2D decomposition of 5D data (M = 2 and N = 5), the dimensions (4, 5) are decomposed by default.

The optional parameter perm should be a (compile-time) tuple defining a permutation of the data indices. Such permutation may be useful for performance reasons, since it may be preferable (e.g. for FFTs) that the data is contiguous along the pencil direction.

It is also possible to pass a TimerOutput to the constructor. See Measuring performance for details.

Examples

Decompose a 3D geometry of global dimensions $N_x × N_y × N_z = 4×8×12$ along the second ($y$) and third ($z$) dimensions.

Pencil(topology, (4, 8, 12), (2, 3))                                # data is in (x, y, z) order
Pencil(topology, (4, 8, 12), (2, 3), permute=Permutation(3, 2, 1))  # data is in (z, y, x) order

In the second case, the actual data is stored in (z, y, x) order within each MPI process.

Pencil(
    p::Pencil{N,M};
    decomp_dims::Dims{M} = decomposition(p),
    size_global::Dims{N} = size_global(p),
    permute::P = permutation(p),
    timer::TimerOutput = timer(p),
)

Create new pencil configuration from an existent one.

This constructor enables sharing temporary data buffers between the two pencil configurations, leading to reduced global memory usage.

PencilArrays.Pencils.AbstractIndexOrder — Type

AbstractIndexOrder

Abstract type determining the ordering of dimensions of an array with possibly permuted indices.

Subtypes are MemoryOrder and LogicalOrder.

PencilArrays.Pencils.MemoryOrder — Type

MemoryOrder <: AbstractIndexOrder

Singleton type specifying that array dimensions should be given in memory (or permuted) order.

PencilArrays.Pencils.LogicalOrder — Type

LogicalOrder <: AbstractIndexOrder

Singleton type specifying that array dimensions should be given in logical (or non-permuted) order.

Methods

PencilArrays.Pencils.topology — Method

topology(p::Pencil)

Get MPITopology attached to Pencil.

PencilArrays.Pencils.MPITopologies.get_comm — Method

get_comm(p::Pencil)

Get MPI communicator associated to an MPI decomposition scheme.

PencilArrays.Pencils.decomposition — Method

decomposition(p::Pencil)

Get tuple with decomposed dimensions of the given pencil configuration.

PencilArrays.Pencils.permutation — Method

permutation(::Type{<:Pencil}) -> AbstractPermutation
permutation(p::Pencil)        -> AbstractPermutation

Get index permutation associated to the given pencil configuration.

Returns NoPermutation() if there is no associated permutation.

PencilArrays.Pencils.timer — Method

timer(p::Pencil)

Get TimerOutput attached to a Pencil.

See Measuring performance for details.

Base.length — Method

length(p::Pencil)

Get linear length of data associated to the local pencil layout.

Base.ndims — Method

ndims(t::MPITopology)

Get dimensionality of Cartesian topology.

ndims(p::Pencil)

Number of spatial dimensions associated to pencil data.

This corresponds to the total number of dimensions of the space, which includes the decomposed and non-decomposed dimensions.

PencilArrays.Pencils.range_remote — Method

range_remote(p::Pencil, coords, [order = LogicalOrder()])
range_remote(p::Pencil, n::Integer, [order = LogicalOrder()])

Get data range held by a given MPI process.

In the first variant, coords are the coordinates of the MPI process in the Cartesian topology. They can be specified as a tuple (i, j, ...) or as a CartesianIndex.

In the second variant, n is the linear index of a given process in the topology.

PencilArrays.Pencils.range_local — Method

range_local(p::Pencil, [order = LogicalOrder()])

Local data range held by the pencil.

By default the dimensions are not permuted, i.e. they follow the logical order of dimensions.

PencilArrays.Pencils.size_global — Method

size_global(p::Pencil, [order = LogicalOrder()])

Global dimensions of the Cartesian grid associated to the given domain decomposition.

Like size_local, by default the returned dimensions are in logical order.

PencilArrays.Pencils.size_local — Method

size_local(p::Pencil, [order = LogicalOrder()])

Local dimensions of the data held by the pencil.

By default the dimensions are not permuted, i.e. they follow the logical order of dimensions.

PencilArrays.Pencils.to_local — Method

to_local(p::Pencil, global_inds, [order = LogicalOrder()])

Convert non-permuted (logical) global indices to local indices.

If order = MemoryOrder(), returned indices will be permuted using the permutation associated to the pencil configuration p.

Index

PencilArrays.Pencils.AbstractIndexOrder

PencilArrays.Pencils.LogicalOrder

PencilArrays.Pencils.MemoryOrder

PencilArrays.Pencils.Pencil

Base.length

Base.ndims

PencilArrays.Pencils.MPITopologies.get_comm

PencilArrays.Pencils.decomposition

PencilArrays.Pencils.permutation

PencilArrays.Pencils.range_local

PencilArrays.Pencils.range_remote

PencilArrays.Pencils.size_global

PencilArrays.Pencils.size_local

PencilArrays.Pencils.timer

PencilArrays.Pencils.to_local

PencilArrays.Pencils.topology

1Why would we want this? One application is to efficiently perform FFTs along $y$, which, under this permutation, would be the fastest dimension. This is used by the PencilFFTs package.