Pencil configurations

A pencil configuration refers to a given distribution of multidimensional data among MPI processes. This information is encoded in the Pencil type.

A pencil configuration includes:

MPI topology information,

global and local dimensions of the numerical grid,

subset of decomposed dimensions,

definition of optional permutation of dimensions.

Construction

The creation of a new Pencil requires a MPITopology, as well as the global data dimensions and a list of decomposed dimensions. Optionally, one can also specify a permutation of dimensions.

For instance, to decompose along the first and third dimensions of a complex 3D dataset,

topology = MPITopology(comm, (8, 4))  # assuming 8×4 = 32 processes
dims_global = (16, 32, 64)
decomp_dims = (1, 3)
pencil = Pencil(topology, dims_global, decomp_dims)

One may also want to create multiple pencil configurations that differ, for instance, on the selection of decomposed dimensions. For this case, a second constructor is available that takes an already existing Pencil instance. Calling this constructor should be preferred when possible since it allows sharing memory buffers (used for instance for global transpositions) and thus reducing memory usage. The following creates a Pencil equivalent to the one above, but with different decomposed dimensions:

pencil_x = Pencil(pencil, decomp_dims=(2, 3))

See the Pencil documentation for more details.

Dimension permutations

As mentioned above, a Pencil may optionally be given information on dimension permutations. In this case, the layout of the data arrays in memory is different from the logical order of dimensions.

To make this clearer, consider the example above where the global data dimensions are $N_x × N_y × N_z = 16 × 32 × 64$. In this case, the logical order is $(x, y, z)$. Now let's say that we want the memory order of the data to be $(y, z, x)$,[1] which corresponds to the permutation (2, 3, 1).

Permutations are passed to the Pencil constructor via the permute keyword argument. For performance reasons, dimension permutations are compile-time constants, and they should be specified using the Permutation type defined in PencilArrays. For instance,

permutation = Permutation(2, 3, 1)
pencil = Pencil(#= ... =#, permute=permutation)

One can also pass NoPermutation as a permutation, which disables permutations (this is the default).

Types

PencilArrays.Pencils.Pencil — Type

Pencil{N,M}

Describes the decomposition of an N-dimensional array among MPI processes along M directions (with M < N).

Pencil(
    topology::MPITopology{M}, size_global::Dims{N}, decomp_dims::Dims{M};
    permute::Permutation = NoPermutation(),
    timer = TimerOutput(),
)

Define the decomposition of an N-dimensional geometry along M dimensions.

The dimensions of the geometry are given by size_global = (N1, N2, ...). The Pencil describes the decomposition of an array of dimensions size_global across a group of MPI processes.

Data is distributed over the given M-dimensional MPI topology (with M < N). The decomposed dimensions are given by decomp_dims.

The optional parameter perm should be a (compile-time) tuple defining a permutation of the data indices. Such permutation may be useful for performance reasons, since it may be preferable (e.g. for FFTs) that the data is contiguous along the pencil direction.

It is also possible to pass a TimerOutput to the constructor. See Measuring performance for details.

Examples

Decompose a 3D geometry of global dimensions $N_x × N_y × N_z = 4×8×12$ along the second ($y$) and third ($z$) dimensions.

Pencil(topology, (4, 8, 12), (2, 3))                                # data is in (x, y, z) order
Pencil(topology, (4, 8, 12), (2, 3), permute=Permutation(3, 2, 1))  # data is in (z, y, x) order

In the second case, the actual data is stored in (z, y, x) order within each MPI process.

Pencil(
    p::Pencil{N,M};
    decomp_dims::Dims{M} = get_decomposition(p),
    size_global::Dims{N} = size_global(p),
    permute::P = get_permutation(p),
    timer::TimerOutput = get_timer(p),
)

Create new pencil configuration from an existent one.

This constructor enables sharing temporary data buffers between the two pencil configurations, leading to reduced global memory usage.

PencilArrays.Permutations.Permutation — Type

Permutation{p}

Describes a compile-time dimension permutation.

The type parameter p should be a valid permutation such as (3, 1, 2).

The parameter p may also be nothing. A Permutation{nothing} represents an identity permutation: it is equivalent to Permutation{(1, 2, …, N)} for any number of dimensions N. The NoPermutation type is provided as an alias for Permutation{nothing}.

Permutation(perm::Vararg{Int})
Permutation(perm::NTuple{N,Int})

Constructs a Permutation.

Example

Both are equivalent:

p1 = Permutation(3, 4)
p2 = Permutation((3, 4))

Permutation(nothing)
NoPermutation()

Constructs an identity permutation.

PencilArrays.Permutations.NoPermutation — Type

NoPermutation

Alias for an identity permutation, i.e. NoPermutation = Permutation{nothing}.

This alias can be called as a constructor: NoPermutation() creates a Permutation{nothing}, as expected.

PencilArrays.Pencils.AbstractIndexOrder — Type

AbstractIndexOrder

Abstract type determining the ordering of dimensions of an array with possibly permuted indices.

Subtypes are MemoryOrder and LogicalOrder.

PencilArrays.Pencils.MemoryOrder — Type

MemoryOrder <: AbstractIndexOrder

Singleton type specifying that array dimensions should be given in memory (or permuted) order.

PencilArrays.Pencils.LogicalOrder — Type

LogicalOrder <: AbstractIndexOrder

Singleton type specifying that array dimensions should be given in logical (or non-permuted) order.

Methods

PencilArrays.Pencils.topology — Method

topology(p::Pencil)

Get MPITopology attached to Pencil.

PencilArrays.Pencils.MPITopologies.get_comm — Method

get_comm(p::Pencil)

Get MPI communicator associated to an MPI decomposition scheme.

PencilArrays.Pencils.get_decomposition — Method

get_decomposition(p::Pencil)

Get tuple with decomposed dimensions of the given pencil configuration.

PencilArrays.Pencils.get_permutation — Method

get_permutation(p::Pencil)

Get index permutation associated to the given pencil configuration.

Returns NoPermutation() if there is no associated permutation.

PencilArrays.Pencils.get_timer — Method

get_timer(p::Pencil)

Get TimerOutput attached to a Pencil.

See Measuring performance for details.

Base.length — Method

length(p::Pencil)

Get linear length of data associated to the local pencil layout.

Base.ndims — Method

ndims(t::MPITopology)

Get dimensionality of Cartesian topology.

ndims(p::Pencil)

Number of spatial dimensions associated to pencil data.

This corresponds to the total number of dimensions of the space, which includes the decomposed and non-decomposed dimensions.

PencilArrays.Pencils.range_remote — Method

range_remote(p::Pencil, coords, [order = LogicalOrder()])
range_remote(p::Pencil, n::Integer, [order = LogicalOrder()])

Get data range held by a given MPI process.

In the first variant, coords are the coordinates of the MPI process in the Cartesian topology. They can be specified as a tuple (i, j, ...) or as a CartesianIndex.

In the second variant, n is the linear index of a given process in the topology.

PencilArrays.Pencils.range_local — Method

range_local(p::Pencil, [order = LogicalOrder()])

Local data range held by the pencil.

By default the dimensions are not permuted, i.e. they follow the logical order of dimensions.

PencilArrays.Pencils.size_global — Method

size_global(p::Pencil, [order = LogicalOrder()])

Global dimensions of the Cartesian grid associated to the given domain decomposition.

Like size_local, by default the returned dimensions are in logical order.

PencilArrays.Pencils.size_local — Method

size_local(p::Pencil, [order = LogicalOrder()])

Local dimensions of the data held by the pencil.

By default the dimensions are not permuted, i.e. they follow the logical order of dimensions.

PencilArrays.Pencils.to_local — Method

to_local(p::Pencil, global_inds, [order = LogicalOrder()])

Convert non-permuted (logical) global indices to local indices.

If order = MemoryOrder(), returned indices will be permuted using the permutation associated to the pencil configuration p.

Index

PencilArrays.Pencils.AbstractIndexOrder

PencilArrays.Pencils.LogicalOrder

PencilArrays.Pencils.MemoryOrder

PencilArrays.Pencils.Pencil

PencilArrays.Permutations.NoPermutation

PencilArrays.Permutations.Permutation

Base.length

Base.ndims

PencilArrays.Pencils.MPITopologies.get_comm

PencilArrays.Pencils.get_decomposition

PencilArrays.Pencils.get_permutation

PencilArrays.Pencils.get_timer

PencilArrays.Pencils.range_local

PencilArrays.Pencils.range_remote

PencilArrays.Pencils.size_global

PencilArrays.Pencils.size_local

PencilArrays.Pencils.to_local

PencilArrays.Pencils.topology

1Why would we want this? Perhaps because we want to efficiently perform FFTs along $y$, which, under this permutation, would be the fastest dimension. This is used by the PencilFFTs package.